+++ ipfs-to-cf-images/.gitignore
@@ -2,3 +2,4 @@
 res/
 __pycache__/
 meta/
+missing/
