+++ ipfs-to-cf-images/random_utils.py
@@ -1,0 +1,4 @@
+def fast_extract(ipfs):
+  if not ipfs:
+    return ''
+  return ipfs.replace('ipfs://ipfs/', '')
